+++ NEWS.md
@@ -20,6 +20,7 @@ form or as a collection, and when printing their summaries. Simplification
 controlled by formal parameter `attr.simplify` with default to `TRUE`.
 * Warn if `find_spikes()` is called with spectral data expressed on irregular 
 wavelength steps, such as after applying `thin_wl()`.
+* Bug fix: in `trimInstrDesc()` handle correctly missing record fields. 
 
 # photobiology 0.12.0